+++ pkg/tsdb/azuremonitor/aztokenprovider/token_cache_test.go
@@ -20,7 +20,7 @@ type fakeRetriever struct {
 	getAccessTokenFunc func(ctx context.Context, scopes []string) (*AccessToken, error)
 }
 
-func (c *fakeRetriever) GetCacheKey() string {
+func (c *fakeRetriever) GetCacheKey(ctx context.Context) string {
 	return c.key
 }
